Added admin functionality, category creation queue, and user permissions. Updated SQL Schema to accomodate

+++ backend/src/main/java/com/statit/backend/dto/ScoreInfoResponse.java
@@ -21,7 +21,7 @@ import java.util.UUID;
 //----------------------------------------------------------------------------------------------------
 public record ScoreInfoResponse(
         UUID scoreId,
-        Float score,
+        Double score,
         String categoryName,
         UUID categoryId,
         String units,
